requirements.txt
streamlit==1.28.0
langchain==0.3.25
langchain-openai==0.3.23
python-dotenv==1.0.0
openai==1.0.0